\name{spear}
\alias{spear}
\title{Calculate SPEAR values}
\usage{
  spear(x, taxa = NULL, group = NULL, abundance = NULL,
    region = NULL, traits = NULL)
}
\arguments{
  \item{x}{Abundance data in long format}

  \item{taxa}{character string: columnnames with taxa}

  \item{group}{character vector naming the columns for
  groups}

  \item{abundance}{character string: columnname of
  abundances}

  \item{region}{character string: specify region}

  \item{traits}{path to traits database}
}
\description{
  Calculate SPEAR values
}

