\name{get_traits}
\alias{get_traits}
\title{Get Trait-data for SPEAR}
\usage{
  get_traits()
}
\description{
  Download trait-data from SPEAR
  (http://www.systemecology.eu/spear/spear-calculator/)
}
\details{
  Trait data is saved as file ('traits.csv') on the local
  machine in order to reduce traffic on server.
}

